refactor: extract type hint handlers to their individual commands & DisplayValues method to ConsoleUtilities
